# userservices/.optic/generated/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Optic Generated OpenAPI",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/hospital/getInvoiceGraphToHospital": {
      "get": {
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "success": {
                          "type": "boolean"
                        },
                        "Totalinvoicescount": {
                          "type": "number"
                        },
                        "TotalcountRejected": {
                          "type": "number"
                        },
                        "TotalcountInProcess": {
                          "type": "number"
                        },
                        "TotalcountDisbursed": {
                          "type": "number"
                        }
                      },
                      "required": [
                        "success",
                        "Totalinvoicescount",
                        "TotalcountRejected",
                        "TotalcountInProcess",
                        "TotalcountDisbursed"
                      ]
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/api/hospital/buildupdatetorepaument": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "_ids": {
                    "oneOf": [
                      {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {}
                        }
                      },
                      {
                        "type": "string"
                      }
                    ]
                  },
                  "ReceivedNEFT_RTG": {
                    "type": "string"
                  },
                  "PaymentReceivedDate": {
                    "type": "string"
                  }
                },
                "required": [
                  "_ids",
                  "ReceivedNEFT_RTG",
                  "PaymentReceivedDate"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "success": {
                          "type": "boolean"
                        },
                        "message": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "success",
                        "message"
                      ]
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ]
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "errors": {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "message"
                      ]
                    }
                  },
                  "required": [
                    "errors"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  }
}